CCP-104 chore/api(all): fix all and improve all systems

+++ cacompilepas/back/App/Module/CourseSheetModule/Model/CourseSheetModel.php
@@ -7,13 +7,15 @@ use App\Http\JSONException;
 use App\Model\AbstractFileModel;
 use App\Model\AbstractModel;
 use App\Module\CourseSheetModule\Exception\InexistentCourseSheetException;
+use App\Module\CourseSheetModule\Exception\InexistentHashtagException;
 use App\Module\LobbyModule\Exception\InexistentLobbyException;
 
 class CourseSheetModel extends AbstractFileModel
 {
+
     public function getLobbyId(int $courseSheetId): int
     {
-        $this->send_query('
+        $this->sendQuery('
             SELECT id_course_sheet
             FROM ccp_coursesheet
             WHERE id_course_sheet = ?
@@ -21,7 +23,7 @@ class CourseSheetModel extends AbstractFileModel
             [$courseSheetId]);
 
         if ($result = $this->getQuery()->fetch()) {
-            $this->send_query('
+            $this->sendQuery('
             SELECT id_lobby_contain
             FROM ccp_coursesheet
             WHERE id_course_sheet = ?
@@ -40,9 +42,9 @@ class CourseSheetModel extends AbstractFileModel
         }
     }
 
-    public function getCourseSheets(int $idLobby): array
+    public function getCourseSheets(int $lobbyId): array
     {
-        $this->send_query('
+        $this->sendQuery('
             SELECT ccp_coursesheet.id_course_sheet, title, publication_date, file_name, ccp_coursesheet.description, pseudo
             FROM ccp_coursesheet 
             INNER JOIN ccp_lobby cl ON ccp_coursesheet.id_lobby_contain = cl.id_lobby
@@ -50,49 +52,52 @@ class CourseSheetModel extends AbstractFileModel
             INNER JOIN ccp_user cu on cia.id_user = cu.id_user
             WHERE id_lobby_Contain = ?
         ',
-            [$idLobby]);
+            [$lobbyId]);
         return $this->fetchData('Lobby does not contain any course sheet');
     }
 
-    public function addCourseSheet(int $idLobby, string $title, string $fileName, string $tmpName, string $description, array $hashtags): array
+    public function addCourseSheet(int $lobbyId, string $title, string $fileName, string $tmpName, string $description, array $hashtags): array
     {
-        $successfulInsert = $this->send_query('
+        $successfulInsert = $this->sendQuery('
             INSERT INTO ccp_coursesheet
             (title, publication_date, file_name, description, id_lobby_contain)
             VALUES
             (?, NOW(), ?, ?, ?)
         ',
-            [$title, $fileName, $description, $idLobby]);
+            [$title, $fileName, $description, $lobbyId]);
 
-        $this->send_query('
+        if ($successfulInsert) {
+
+            $this->sendQuery('
             SELECT id_course_sheet
             FROM ccp_coursesheet
             ORDER BY id_course_sheet DESC
             LIMIT 1
         ',
-            []);
+                []);
 
-        $courseSheetId = (int)$this->getQuery()->fetch()[0];
+            $courseSheetId = (int)$this->fetchData('Course sheet does not exist')['data'][0]['id_course_sheet'];
 
-        try {
-            $this->uploadOnFTP($idLobby, $fileName, $tmpName, '/coursesheets/', AbstractModel::$COURSE_SHEET_EXTENSIONS);
-        } catch (IncorrectFileExtension $e) {
-            throw new JSONException($e->getMessage());
-        }
+            try {
+                $this->uploadOnFTP($courseSheetId, $fileName, $tmpName, '/course_sheets/', AbstractModel::$COURSE_SHEET_EXTENSIONS);
+            } catch (IncorrectFileExtension $e) {
+                throw new JSONException($e->getMessage());
+            } catch (JSONException $e) {
+                throw $e;
+            }
 
-        foreach ($hashtags as $key => $value) {
-            if ($successfulInsert && $successfulUpload) {
-                $successfulInsert = $this->send_query('
+            foreach ($hashtags as $key => $value) {
+                if ($successfulInsert && $successfulUpload) {
+                    $successfulInsert = $this->sendQuery('
                 INSERT INTO ccp_hashtag
                 (label_hashtag, id_course_sheet) 
                 VALUES 
                 (?, ?)
             ',
-                    [$value, $courseSheetId]);
+                        [$value, $courseSheetId]);
+                }
             }
-        }
 
-        if ($successfulInsert && $successfulUpload) {
             return [
                 'message' => 'Course sheet was successfully added',
             ];
@@ -101,9 +106,9 @@ class CourseSheetModel extends AbstractFileModel
         }
     }
 
-    public function deleteCourseSheet(int $idLobby, $courseSheetId): array
+    public function deleteCourseSheet(int $lobbyId, $courseSheetId): array
     {
-        $this->send_query('
+        $this->sendQuery('
                 SELECT file_name
                 FROM ccp_coursesheet
                 WHERE id_course_sheet = ?
@@ -111,22 +116,22 @@ class CourseSheetModel extends AbstractFileModel
             [$courseSheetId]);
         $fileName = $this->getQuery()->fetch()['file_name'];
 
-        $successfulDelete = $this->send_query('
+        $successfulDelete = $this->sendQuery('
             DELETE FROM ccp_coursesheet
             WHERE id_lobby_contain = ?
             AND id_course_sheet = ?
         ',
-            [$idLobby, $courseSheetId]);
+            [$lobbyId, $courseSheetId]);
         if ($successfulDelete) {
-            $this->send_query('
+            $this->sendQuery('
                 SELECT file_name
                 FROM ccp_coursesheet
                 WHERE id_course_sheet = ?
             ',
                 [$courseSheetId]);
 
-            $oldFileNameOnFTPServer = '/coursesheets/' . $this->nameOnFTP($courseSheetId, $fileName, $this->extension($fileName));
-            $this->deleteOnFTP($oldFileNameOnFTPServer, '/coursesheets/');
+            $oldFileNameOnFTPServer = '/course_sheets/' . $this->nameOnFTP($courseSheetId, $fileName, $this->extension($fileName));
+            $this->deleteOnFTP($oldFileNameOnFTPServer, '/course_sheets/');
 
             return [
                 'message' => 'Course sheet was successfully deleted',
@@ -139,13 +144,13 @@ class CourseSheetModel extends AbstractFileModel
     public function addHashtags(int $courseSheetId, array $hashtags): array
     {
         foreach ($hashtags as $key => $value) {
-            $successfullyAdded = $this->send_query("
+            $successfullyAdded = $this->sendQuery("
                 INSERT INTO ccp_hashtag
                 (label_hashtag, id_course_sheet) VALUES (?, ?)
             ",
                 [$value, $courseSheetId]);
             if (!$successfullyAdded) {
-                throw new JSONException('Hashtags could not be added');
+                throw new JSONException('One or more of hashtags is already associated with that course sheet');
             }
         }
         return [
@@ -155,24 +160,37 @@ class CourseSheetModel extends AbstractFileModel
 
     public function removeHashtag(int $courseSheetId, string $hashtag): array
     {
-        $successfullyRemoved = $this->send_query('
+        $this->sendQuery('
+            SELECT label_hashtag
+            FROM ccp_hashtag
+            INNER JOIN ccp_coursesheet cc ON ccp_hashtag.id_course_sheet = cc.id_course_sheet
+            WHERE cc.id_course_sheet = ?
+            And label_hashtag = ?
+        ',
+            [$courseSheetId, $hashtag]);
+
+        if ($result = $this->getQuery()->fetch()) {
+            $successfullyRemoved = $this->sendQuery('
             DELETE FROM ccp_hashtag
             WHERE label_hashtag = ?
             AND id_course_sheet = ?
         ',
-            [$hashtag, $courseSheetId]);
-        if (!$successfullyRemoved) {
-            throw new JSONException('Hashtag could not be removed');
+                [$hashtag, $courseSheetId]);
+            if (!$successfullyRemoved) {
+                throw new JSONException('Hashtag could not be removed');
+            } else {
+                return [
+                    'message' => 'Hashtag was successfully removed',
+                ];
+            }
         } else {
-            return [
-                'message' => 'Hashtag was successfully removed',
-            ];
+            throw new InexistentHashtagException();
         }
     }
 
     public function getHashtags(int $courseSheetId): array
     {
-        $this->send_query('
+        $this->sendQuery('
             SELECT label_hashtag
             FROM ccp_hashtag
             INNER JOIN ccp_coursesheet cc ON ccp_hashtag.id_course_sheet = cc.id_course_sheet
@@ -185,7 +203,7 @@ class CourseSheetModel extends AbstractFileModel
 
     public function getPath(int $courseSheetId): string
     {
-        $this->send_query('
+        $this->sendQuery('
             SELECT file_name FROM ccp_coursesheet
             WHERE id_course_sheet = ?
         ',
